# contracts/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/assets":{"get":{"tags":["capability-asset-controller"],"operationId":"search","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["DRAFT","IN_REVIEW","APPROVED","REJECTED","DEPRECATED"]}},{"name":"assetType","in":"query","required":false,"schema":{"type":"string","enum":["PROMPT_TEMPLATE","WORKFLOW_AUTOMATION","AI_AGENT","KNOWLEDGE_BOT","ANALYTICS_BRIEF","CONTENT_GENERATOR","DATA_EXTRACTION","EVALUATION_CHECKLIST","PLAYBOOK","HANDOVER_PACK","GOVERNANCE_GUARDRAIL","COPILOT"]}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"post":{"tags":["capability-asset-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/usage":{"post":{"tags":["capability-asset-controller"],"operationId":"recordUsage","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsageResponse"}}}}}}},"/api/ai/chat":{"post":{"tags":["asset-assistant-controller"],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/SseEmitter"}}}}}}},"/api/ai/assets/normalize":{"post":{"tags":["asset-assistant-controller"],"operationId":"normalize","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiDraftRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AiDraftResponse"}}}}}}},"/api/assets/{assetId}/submit-review":{"patch":{"tags":["capability-asset-controller"],"operationId":"submitForReview","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/reject":{"patch":{"tags":["capability-asset-controller"],"operationId":"reject","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/deprecate":{"patch":{"tags":["capability-asset-controller"],"operationId":"deprecate","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/backup-owner":{"patch":{"tags":["capability-asset-controller"],"operationId":"assignBackupOwner","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BackupOwnerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/approve":{"patch":{"tags":["capability-asset-controller"],"operationId":"approve","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/session":{"get":{"tags":["browser-session-controller"],"summary":"Read the current browser session","operationId":"getBrowserSession","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/api/session/csrf":{"get":{"tags":["browser-session-controller"],"summary":"Issue a CSRF token for browser mutations","operationId":"getBrowserCsrfToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CsrfResponse"}}}}}}},"/api/organization/context":{"get":{"tags":["organization-context-controller"],"operationId":"context","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationContextResponse"}}}}}}},"/api/me":{"get":{"tags":["me-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}}},"/api/knowledge-assets":{"get":{"tags":["knowledge-asset-controller"],"operationId":"search_1","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KnowledgeAssetSummaryResponse"}}}}}}}},"/api/knowledge-assets/{assetId}":{"get":{"tags":["knowledge-asset-controller"],"operationId":"get","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/KnowledgeAssetDetailResponse"}}}}}}},"/api/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/graph":{"get":{"tags":["knowledge-graph-controller"],"operationId":"graph","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"focusAssetId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"depth","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":2}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/KnowledgeGraphResponse"}}}}}}},"/api/assets/{assetId}":{"get":{"tags":["capability-asset-controller"],"operationId":"get_1","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CapabilityAssetResponse"}}}}}}},"/api/assets/{assetId}/versions":{"get":{"tags":["capability-asset-controller"],"operationId":"versions","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssetVersionResponse"}}}}}}}}},"components":{"schemas":{"CreateAssetRequest":{"type":"object","properties":{"organizationId":{"type":"string","format":"uuid"},"departmentId":{"type":"string","format":"uuid"},"title":{"type":"string","minLength":1},"summary":{"type":"string","minLength":1},"assetType":{"type":"string","enum":["PROMPT_TEMPLATE","WORKFLOW_AUTOMATION","AI_AGENT","KNOWLEDGE_BOT","ANALYTICS_BRIEF","CONTENT_GENERATOR","DATA_EXTRACTION","EVALUATION_CHECKLIST","PLAYBOOK","HANDOVER_PACK","GOVERNANCE_GUARDRAIL","COPILOT"]},"useCase":{"type":"string"},"businessProcess":{"type":"string"},"aiTool":{"type":"string"},"tagNames":{"type":"string"},"ownerUserId":{"type":"string","format":"uuid"},"backupOwnerUserId":{"type":"string","format":"uuid"},"createdByUserId":{"type":"string","format":"uuid"},"visibility":{"type":"string","enum":["PRIVATE","TEAM","ORGANIZATION"]},"riskLevel":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"promptTemplate":{"type":"string"},"workflowStepsJson":{"type":"string"},"inputSchemaJson":{"type":"string"},"outputSchemaJson":{"type":"string"},"exampleInput":{"type":"string"},"exampleOutput":{"type":"string"}},"required":["summary","title"]},"CapabilityAssetResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"departmentId":{"type":"string","format":"uuid"},"title":{"type":"string"},"summary":{"type":"string"},"assetType":{"type":"string","enum":["PROMPT_TEMPLATE","WORKFLOW_AUTOMATION","AI_AGENT","KNOWLEDGE_BOT","ANALYTICS_BRIEF","CONTENT_GENERATOR","DATA_EXTRACTION","EVALUATION_CHECKLIST","PLAYBOOK","HANDOVER_PACK","GOVERNANCE_GUARDRAIL","COPILOT"]},"useCase":{"type":"string"},"businessProcess":{"type":"string"},"aiTool":{"type":"string"},"tagNames":{"type":"string"},"ownerUserId":{"type":"string","format":"uuid"},"backupOwnerUserId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["DRAFT","IN_REVIEW","APPROVED","REJECTED","DEPRECATED"]},"visibility":{"type":"string","enum":["PRIVATE","TEAM","ORGANIZATION"]},"riskLevel":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"currentVersionId":{"type":"string","format":"uuid"},"createdByUserId":{"type":"string","format":"uuid"},"usageCount":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UsageRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"eventType":{"type":"string","enum":["VIEWED","COPIED","USED","SHARED"]},"metadataJson":{"type":"string"}},"required":["eventType"]},"UsageResponse":{"type":"object","properties":{"assetId":{"type":"string","format":"uuid"},"usageCount":{"type":"integer","format":"int64"}}},"ChatRequest":{"type":"object","properties":{"message":{"type":"string"},"conversationId":{"type":"string"}}},"SseEmitter":{"type":"object","properties":{"timeout":{"type":"integer","format":"int64"}}},"AiDraftRequest":{"type":"object","properties":{"rawText":{"type":"string","minLength":1},"aiTool":{"type":"string"},"businessProcess":{"type":"string"}},"required":["rawText"]},"AiDraftResponse":{"type":"object","properties":{"aiEnabled":{"type":"boolean"},"source":{"type":"string"},"note":{"type":"string"},"title":{"type":"string"},"summary":{"type":"string"},"assetType":{"type":"string","enum":["PROMPT_TEMPLATE","WORKFLOW_AUTOMATION","AI_AGENT","KNOWLEDGE_BOT","ANALYTICS_BRIEF","CONTENT_GENERATOR","DATA_EXTRACTION","EVALUATION_CHECKLIST","PLAYBOOK","HANDOVER_PACK","GOVERNANCE_GUARDRAIL","COPILOT"]},"useCase":{"type":"string"},"businessProcess":{"type":"string"},"aiTool":{"type":"string"},"tagNames":{"type":"string"},"riskLevel":{"type":"string","enum":["LOW","MEDIUM","HIGH"]},"promptTemplate":{"type":"string"},"workflowStepsJson":{"type":"string"},"inputSchemaJson":{"type":"string"},"outputSchemaJson":{"type":"string"},"exampleInput":{"type":"string"},"exampleOutput":{"type":"string"}}},"ReviewRequest":{"type":"object","properties":{"reviewerUserId":{"type":"string","format":"uuid"},"comment":{"type":"string"}}},"BackupOwnerRequest":{"type":"object","properties":{"backupOwnerUserId":{"type":"string","format":"uuid"}},"required":["backupOwnerUserId"]},"SessionResponse":{"type":"object","properties":{"authenticated":{"type":"boolean"},"name":{"type":"string"},"email":{"type":"string"},"userId":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"departmentId":{"type":"string","format":"uuid"}}},"CsrfResponse":{"type":"object","properties":{"headerName":{"type":"string"},"parameterName":{"type":"string"},"token":{"type":"string"}}},"DepartmentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"OrganizationContextResponse":{"type":"object","properties":{"organizationId":{"type":"string","format":"uuid"},"departments":{"type":"array","items":{"$ref":"#/components/schemas/DepartmentResponse"}},"users":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"departmentId":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["EMPLOYEE","TEAM_LEAD","MANAGER","DIRECTOR","EXECUTIVE","ADMIN"]}}},"MeResponse":{"type":"object","properties":{"authenticated":{"type":"boolean"},"subject":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"authorizationProvider":{"type":"string"},"userId":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"departmentId":{"type":"string","format":"uuid"}}},"KnowledgeAssetSummaryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"departmentId":{"type":"string","format":"uuid"},"classification":{"type":"string","enum":["PUBLIC","INTERNAL","CONFIDENTIAL","RESTRICTED"]},"updatedAt":{"type":"string","format":"date-time"}}},"KnowledgeAssetDetailResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"content":{"type":"string"},"language":{"type":"string"},"departmentId":{"type":"string","format":"uuid"},"classification":{"type":"string","enum":["PUBLIC","INTERNAL","CONFIDENTIAL","RESTRICTED"]},"sourceSystem":{"type":"string"},"externalObjectId":{"type":"string"},"sourceUri":{"type":"string"},"activatedAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"KnowledgeGraphEdgeResponse":{"type":"object","properties":{"id":{"type":"string"},"source":{"type":"string"},"target":{"type":"string"},"kind":{"type":"string","enum":["HAS_TYPE","BELONGS_TO","OWNED_BY","BACKED_UP_BY","TAGGED_WITH","RELATED_BY_TAG","RELATED_BY_OWNER","RELATED_BY_PROCESS"]},"label":{"type":"string"},"weight":{"type":"integer","format":"int32"}}},"KnowledgeGraphNodeResponse":{"type":"object","properties":{"id":{"type":"string"},"label":{"type":"string"},"kind":{"type":"string","enum":["ASSET","ASSET_TYPE","DEPARTMENT","USER","TAG"]},"detail":{"type":"string"},"assetId":{"type":"string"},"assetType":{"type":"string"},"status":{"type":"string"},"weight":{"type":"integer","format":"int32"}}},"KnowledgeGraphResponse":{"type":"object","properties":{"nodes":{"type":"array","items":{"$ref":"#/components/schemas/KnowledgeGraphNodeResponse"}},"edges":{"type":"array","items":{"$ref":"#/components/schemas/KnowledgeGraphEdgeResponse"}},"focusNodeId":{"type":"string"},"depth":{"type":"integer","format":"int32"}}},"AssetVersionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"assetId":{"type":"string","format":"uuid"},"versionNumber":{"type":"integer","format":"int32"},"promptTemplate":{"type":"string"},"workflowStepsJson":{"type":"string"},"inputSchemaJson":{"type":"string"},"outputSchemaJson":{"type":"string"},"exampleInput":{"type":"string"},"exampleOutput":{"type":"string"},"changeNote":{"type":"string"},"createdByUserId":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"}}}}}}